html/semantics/disabled-elements/disable.button-setter.htm

<!DOCTYPE html>
<title>HTMLAnchorElement.text setting</title>
<link rel="author" title="Tim Denney" href="mailto:timdenney@gmail.com">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/disabled-elements.html#disabled-elements">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>

<button id="test_btn" disabled="disabled" onclick="set_value(this)" >Button</button>

<button id="test_bool_attr1" disabled="" >Button</button>
<button id="test_bool_attr2" disabled>Button</button>
<button id="test_bool_attr3" disabled="false" >Button</button>
<button id="test_bool_attr4" disabled=true >Button</button>
<button id="test_bool_attr5" disabled=false >Button</button>

<script>

function set_value(e){
  var btn=document.getElementById(e.id);
  btn.setAttribute("check","fired");
}

function eventFire(item, type){
  if (item!=undefined){
    var evObj = document.createEvent('Events');
    evObj.initEvent('click', true, false);
    item.dispatchEvent(evObj);
  }
}

test(function(){
  var btn=document.getElementById("test_btn");
  assert_equals(btn.getAttribute("disabled"),'disabled');

},'Disable Attribute Exist');

test(function(){
  var btn=document.getElementById("test_btn");
  eventFire(btn,'click');
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Click Event Triggers on Disabled Element');


test(function(){
  var btn=document.getElementById("test_bool_attr1");
  eventFire(btn,'click');
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute Empty Quotes');

test(function(){
  var btn=document.getElementById("test_bool_attr2");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute No Value');

test(function(){
  var btn=document.getElementById("test_bool_attr3");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute False Value');

test(function(){
  var btn=document.getElementById("test_bool_attr4");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute No Quotes Value (True)');


test(function(){
  var btn=document.getElementById("test_bool_attr5");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute No Quotes Value (False)');


</script>
